requirements.txt
transformers==4.39.3
pillow==10.3.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
opencv-python
mediapipe==0.10.9
pandas==2.2.2
openai-whisper 
pyaudio